add navdeep ui
