+++ package.json
@@ -10,6 +10,10 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.14.4",
+    "@mui/material": "^5.14.5",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
   },
