+++ src/components/Audio/Audio.css
@@ -1,0 +1,26 @@
+.main {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    height: 100%;
+    width: 100%;
+    gap: 1rem;
+}
+
+.main .status {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 0.5rem;
+    background-color: var(--bg-color-2);
+    width: 100px;
+    height: 100px;
+    border-radius: 50%;
+}
+
+.speaking {
+    border: 2px solid green;
+    box-shadow: 0 0 5px 5px green;
+}
